# backend/openapi.yaml
openapi: 3.0.3
info:
  title: KAYAD API
  description: |
    KAYAD - Kenya's Premium Car Marketplace API
    
    This API provides endpoints for managing car listings, auctions, bids, escrow transactions, and user accounts.
    
    ## Authentication
    
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    
    API requests are rate limited. Standard limits are 100 requests per minute per user.
    
    ## Versioning
    
    The API is versioned using URL path versioning (e.g., /api/v1/cars). Current version is v1.
    
    ## Deprecation Policy
    
    Deprecated endpoints will be supported for at least 6 months before removal.
  version: 1.0.0
  contact:
    name: KAYAD API Support
    email: api@kayad.space
    url: https://kayad.space
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kayad.space/api/v1
    description: Production server
  - url: https://staging-api.kayad.space/api/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Development server
...
paths:
...
  /auctions/{auctionId}/bids:
    get:
      tags:
        - Bids
      summary: List auction bids
      description: Get all bids for an auction
      operationId: listAuctionBids
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: string
          description: Auction ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidListResponse'
        '404':
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 50

    post:
      tags:
        - Bids
      summary: Place bid
      description: Place a bid on an auction
      operationId: placeBid
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: string
          description: Auction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBidRequest'
      responses:
        '201':
          description: Bid placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - auction not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 30
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
    Bid:
      type: object
      required:
        - id
        - auctionId
        - userId
        - amount
        - timestamp
      properties:
        id:
          type: string
          description: Bid ID
        auctionId:
          type: string
          description: Auction ID
        userId:
          type: string
          description: User ID
        amount:
          type: number
          description: Bid amount in KES
        timestamp:
          type: string
          format: date-time
          description: Bid timestamp
    PlaceBidRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          description: Bid amount in KES
    BidResponse:
      type: object
      required:
        - bid
      properties:
        bid:
          $ref: '#/components/schemas/Bid'
    BidListResponse:
      type: object
      required:
        - bids
        - pagination
      properties:
        bids:
          type: array
          items:
            $ref: '#/components/schemas/Bid'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
...